requirements.txt
Flask~=3.0.3
pymongo~=4.10.1
werkzeug~=3.0.4
azure-servicebus
azure-core~=1.32.0
pip~=24.2
wheel~=0.41.2
pytest